feat: proving context
