+++ examples/minroot.rs
@@ -240,7 +240,7 @@ fn main() {
     // verify the recursive SNARK
     println!("Verifying a RecursiveSNARK...");
     let start = Instant::now();
-    let res = recursive_snark.verify(&pp, num_steps, &z0_primary, &z0_secondary);
+    let res = recursive_snark.verify(&pp);
     println!(
       "RecursiveSNARK::verify: {:?}, took {:?}",
       res.is_ok(),
