+++ src/lib.rs
@@ -36,15 +36,13 @@ use errors::NovaError;
 use ff::Field;
 use gadgets::utils::scalar_as_base;
 use nifs::NIFS;
-use prover::ProverKey;
+use prover::{ProverKey, ProvingContext};
 use public_params::PublicParams;
-use r1cs::{R1CSInstance, R1CSWitness, RelaxedR1CSInstance, RelaxedR1CSWitness};
+use r1cs::{R1CSInstance, RelaxedR1CSInstance, RelaxedR1CSWitness};
 use serde::{Deserialize, Serialize};
 use traits::{
-  circuit::StepCircuit,
-  commitment::{Commitment, CommitmentTrait},
-  snark::RelaxedR1CSSNARKTrait,
-  AbsorbInROTrait, Engine, ROConstants, ROTrait,
+  circuit::StepCircuit, commitment::Commitment, snark::RelaxedR1CSSNARKTrait, AbsorbInROTrait,
+  Engine, ROConstants, ROTrait,
 };
 
 /// A SNARK that proves the correct execution of an incremental computation
@@ -57,18 +55,7 @@ where
   C1: StepCircuit<E1::Scalar>,
   C2: StepCircuit<E2::Scalar>,
 {
-  z0_primary: Vec<E1::Scalar>,
-  z0_secondary: Vec<E2::Scalar>,
-  r_W_primary: RelaxedR1CSWitness<E1>,
-  r_U_primary: RelaxedR1CSInstance<E1>,
-  r_W_secondary: RelaxedR1CSWitness<E2>,
-  r_U_secondary: RelaxedR1CSInstance<E2>,
-  l_w_secondary: R1CSWitness<E2>,
-  l_u_secondary: R1CSInstance<E2>,
-  i: usize,
-  zi_primary: Vec<E1::Scalar>,
-  zi_secondary: Vec<E2::Scalar>,
-  _p: PhantomData<(C1, C2)>,
+  ctx_proving: ProvingContext<E1, E2, C1, C2>,
 }
 
 impl<E1, E2, C1, C2> RecursiveSNARK<E1, E2, C1, C2>
@@ -175,18 +162,20 @@ where
       .expect("Nova error synthesis");
 
     Ok(Self {
-      z0_primary: z0_primary.to_vec(),
-      z0_secondary: z0_secondary.to_vec(),
-      r_W_primary,
-      r_U_primary,
-      r_W_secondary,
-      r_U_secondary,
-      l_w_secondary,
-      l_u_secondary,
-      i: 0,
-      zi_primary,
-      zi_secondary,
-      _p: Default::default(),
+      ctx_proving: ProvingContext {
+        z0_primary: z0_primary.to_vec(),
+        z0_secondary: z0_secondary.to_vec(),
+        r_W_primary,
+        r_U_primary,
+        r_W_secondary,
+        r_U_secondary,
+        l_w_secondary,
+        l_u_secondary,
+        i: 0,
+        zi_primary,
+        zi_secondary,
+        _p: Default::default(),
+      },
     })
   }
 
@@ -198,217 +187,15 @@ where
     c_primary: &C1,
     c_secondary: &C2,
   ) -> Result<(), NovaError> {
-    // first step was already done in the constructor
-    if self.i == 0 {
-      self.i = 1;
-      return Ok(());
-    }
-
-    // fold the secondary circuit's instance
-    let (nifs_secondary, (r_U_secondary, r_W_secondary)) = NIFS::prove(
-      &pp.ck_secondary,
-      &pp.ro_consts_secondary,
-      &scalar_as_base::<E1>(pp.digest()),
-      &pp.r1cs_shape_secondary,
-      &self.r_U_secondary,
-      &self.r_W_secondary,
-      &self.l_u_secondary,
-      &self.l_w_secondary,
-    )
-    .expect("Unable to fold secondary");
-
-    let mut cs_primary = SatisfyingAssignment::<E1>::new();
-    let inputs_primary: NovaAugmentedCircuitInputs<E2> = NovaAugmentedCircuitInputs::new(
-      scalar_as_base::<E1>(pp.digest()),
-      E1::Scalar::from(self.i as u64),
-      self.z0_primary.to_vec(),
-      Some(self.zi_primary.clone()),
-      Some(self.r_U_secondary.clone()),
-      Some(self.l_u_secondary.clone()),
-      Some(Commitment::<E2>::decompress(&nifs_secondary.comm_T)?),
-    );
-
-    let circuit_primary: NovaAugmentedCircuit<'_, E2, C1> = NovaAugmentedCircuit::new(
-      &pp.augmented_circuit_params_primary,
-      Some(inputs_primary),
-      c_primary,
-      pp.ro_consts_circuit_primary.clone(),
-    );
-    let zi_primary = circuit_primary
-      .synthesize(&mut cs_primary)
-      .map_err(|_| NovaError::SynthesisError)?;
-
-    let (l_u_primary, l_w_primary) = cs_primary
-      .r1cs_instance_and_witness(&pp.r1cs_shape_primary, &pp.ck_primary)
-      .map_err(|_e| NovaError::UnSat)
-      .expect("Nova error unsat");
-
-    // fold the primary circuit's instance
-    let (nifs_primary, (r_U_primary, r_W_primary)) = NIFS::prove(
-      &pp.ck_primary,
-      &pp.ro_consts_primary,
-      &pp.digest(),
-      &pp.r1cs_shape_primary,
-      &self.r_U_primary,
-      &self.r_W_primary,
-      &l_u_primary,
-      &l_w_primary,
-    )
-    .expect("Unable to fold primary");
-
-    let mut cs_secondary = SatisfyingAssignment::<E2>::new();
-    let inputs_secondary: NovaAugmentedCircuitInputs<E1> = NovaAugmentedCircuitInputs::new(
-      pp.digest(),
-      E2::Scalar::from(self.i as u64),
-      self.z0_secondary.to_vec(),
-      Some(self.zi_secondary.clone()),
-      Some(self.r_U_primary.clone()),
-      Some(l_u_primary),
-      Some(Commitment::<E1>::decompress(&nifs_primary.comm_T)?),
-    );
-
-    let circuit_secondary: NovaAugmentedCircuit<'_, E1, C2> = NovaAugmentedCircuit::new(
-      &pp.augmented_circuit_params_secondary,
-      Some(inputs_secondary),
-      c_secondary,
-      pp.ro_consts_circuit_secondary.clone(),
-    );
-    let zi_secondary = circuit_secondary
-      .synthesize(&mut cs_secondary)
-      .map_err(|_| NovaError::SynthesisError)?;
-
-    let (l_u_secondary, l_w_secondary) = cs_secondary
-      .r1cs_instance_and_witness(&pp.r1cs_shape_secondary, &pp.ck_secondary)
-      .map_err(|_e| NovaError::UnSat)?;
-
-    // update the running instances and witnesses
-    self.zi_primary = zi_primary
-      .iter()
-      .map(|v| v.get_value().ok_or(NovaError::SynthesisError))
-      .collect::<Result<Vec<<E1 as Engine>::Scalar>, NovaError>>()?;
-    self.zi_secondary = zi_secondary
-      .iter()
-      .map(|v| v.get_value().ok_or(NovaError::SynthesisError))
-      .collect::<Result<Vec<<E2 as Engine>::Scalar>, NovaError>>()?;
-
-    self.l_u_secondary = l_u_secondary;
-    self.l_w_secondary = l_w_secondary;
-
-    self.r_U_primary = r_U_primary;
-    self.r_W_primary = r_W_primary;
-
-    self.i += 1;
-
-    self.r_U_secondary = r_U_secondary;
-    self.r_W_secondary = r_W_secondary;
-
-    Ok(())
+    self.ctx_proving.prove(pp, c_primary, c_secondary)
   }
 
   /// Verify the correctness of the `RecursiveSNARK`
   pub fn verify(
     &self,
     pp: &PublicParams<E1, E2, C1, C2>,
-    num_steps: usize,
-    z0_primary: &[E1::Scalar],
-    z0_secondary: &[E2::Scalar],
   ) -> Result<(Vec<E1::Scalar>, Vec<E2::Scalar>), NovaError> {
-    // number of steps cannot be zero
-    let is_num_steps_zero = num_steps == 0;
-
-    // check if the provided proof has executed num_steps
-    let is_num_steps_not_match = self.i != num_steps;
-
-    // check if the initial inputs match
-    let is_inputs_not_match = self.z0_primary != z0_primary || self.z0_secondary != z0_secondary;
-
-    // check if the (relaxed) R1CS instances have two public outputs
-    let is_instance_has_two_outpus = self.l_u_secondary.X.len() != 2
-      || self.r_U_primary.X.len() != 2
-      || self.r_U_secondary.X.len() != 2;
-
-    if is_num_steps_zero
-      || is_num_steps_not_match
-      || is_inputs_not_match
-      || is_instance_has_two_outpus
-    {
-      return Err(NovaError::ProofVerifyError);
-    }
-
-    // check if the output hashes in R1CS instances point to the right running instances
-    let (hash_primary, hash_secondary) = {
-      let mut hasher = <E2 as Engine>::RO::new(
-        pp.ro_consts_secondary.clone(),
-        NUM_FE_WITHOUT_IO_FOR_CRHF + 2 * pp.F_arity_primary,
-      );
-      hasher.absorb(pp.digest());
-      hasher.absorb(E1::Scalar::from(num_steps as u64));
-      for e in z0_primary {
-        hasher.absorb(*e);
-      }
-      for e in &self.zi_primary {
-        hasher.absorb(*e);
-      }
-      self.r_U_secondary.absorb_in_ro(&mut hasher);
-
-      let mut hasher2 = <E1 as Engine>::RO::new(
-        pp.ro_consts_primary.clone(),
-        NUM_FE_WITHOUT_IO_FOR_CRHF + 2 * pp.F_arity_secondary,
-      );
-      hasher2.absorb(scalar_as_base::<E1>(pp.digest()));
-      hasher2.absorb(E2::Scalar::from(num_steps as u64));
-      for e in z0_secondary {
-        hasher2.absorb(*e);
-      }
-      for e in &self.zi_secondary {
-        hasher2.absorb(*e);
-      }
-      self.r_U_primary.absorb_in_ro(&mut hasher2);
-
-      (
-        hasher.squeeze(NUM_HASH_BITS),
-        hasher2.squeeze(NUM_HASH_BITS),
-      )
-    };
-
-    if hash_primary != self.l_u_secondary.X[0]
-      || hash_secondary != scalar_as_base::<E2>(self.l_u_secondary.X[1])
-    {
-      return Err(NovaError::ProofVerifyError);
-    }
-
-    // check the satisfiability of the provided instances
-    let (res_r_primary, (res_r_secondary, res_l_secondary)) = rayon::join(
-      || {
-        pp.r1cs_shape_primary
-          .is_sat_relaxed(&pp.ck_primary, &self.r_U_primary, &self.r_W_primary)
-      },
-      || {
-        rayon::join(
-          || {
-            pp.r1cs_shape_secondary.is_sat_relaxed(
-              &pp.ck_secondary,
-              &self.r_U_secondary,
-              &self.r_W_secondary,
-            )
-          },
-          || {
-            pp.r1cs_shape_secondary.is_sat(
-              &pp.ck_secondary,
-              &self.l_u_secondary,
-              &self.l_w_secondary,
-            )
-          },
-        )
-      },
-    );
-
-    // check the returned res objects
-    res_r_primary?;
-    res_r_secondary?;
-    res_l_secondary?;
-
-    Ok((self.zi_primary.clone(), self.zi_secondary.clone()))
+    self.ctx_proving.verify(pp)
   }
 }
 
@@ -508,16 +295,30 @@ where
     pk: &ProverKey<E1, E2, C1, C2, S1, S2>,
     recursive_snark: &RecursiveSNARK<E1, E2, C1, C2>,
   ) -> Result<Self, NovaError> {
+    let ProvingContext {
+      z0_primary: _,
+      z0_secondary: _,
+      r_W_primary,
+      r_U_primary,
+      r_W_secondary,
+      r_U_secondary,
+      l_w_secondary,
+      l_u_secondary,
+      i: _,
+      zi_primary,
+      zi_secondary,
+      _p,
+    } = recursive_snark.ctx_proving.clone();
     // fold the secondary circuit's instance with its running instance
     let (nifs_secondary, (f_U_secondary, f_W_secondary)) = NIFS::prove(
       &pp.ck_secondary,
       &pp.ro_consts_secondary,
       &scalar_as_base::<E1>(pp.digest()),
       &pp.r1cs_shape_secondary,
-      &recursive_snark.r_U_secondary,
-      &recursive_snark.r_W_secondary,
-      &recursive_snark.l_u_secondary,
-      &recursive_snark.l_w_secondary,
+      &r_U_secondary,
+      &r_W_secondary,
+      &l_u_secondary,
+      &l_w_secondary,
     )?;
 
     // create SNARKs proving the knowledge of f_W_primary and f_W_secondary
@@ -527,8 +328,8 @@ where
           &pp.ck_primary,
           &pk.pk_primary,
           &pp.r1cs_shape_primary,
-          &recursive_snark.r_U_primary,
-          &recursive_snark.r_W_primary,
+          &r_U_primary,
+          &r_W_primary,
         )
       },
       || {
@@ -543,16 +344,16 @@ where
     );
 
     Ok(Self {
-      r_U_primary: recursive_snark.r_U_primary.clone(),
+      r_U_primary: r_U_primary.clone(),
       r_W_snark_primary: r_W_snark_primary?,
 
-      r_U_secondary: recursive_snark.r_U_secondary.clone(),
-      l_u_secondary: recursive_snark.l_u_secondary.clone(),
+      r_U_secondary: r_U_secondary.clone(),
+      l_u_secondary: l_u_secondary.clone(),
       nifs_secondary,
       f_W_snark_secondary: f_W_snark_secondary?,
 
-      zn_primary: recursive_snark.zi_primary.clone(),
-      zn_secondary: recursive_snark.zi_secondary.clone(),
+      zn_primary: zi_primary.clone(),
+      zn_secondary: zi_secondary.clone(),
 
       _p: Default::default(),
     })
@@ -831,8 +632,6 @@ mod tests {
       &*default_ck_hint(),
     );
 
-    let num_steps = 1;
-
     // produce a recursive SNARK
     let mut recursive_snark = RecursiveSNARK::new(
       &pp,
@@ -848,12 +647,7 @@ mod tests {
     assert!(res.is_ok());
 
     // verify the recursive SNARK
-    let res = recursive_snark.verify(
-      &pp,
-      num_steps,
-      &[<E1 as Engine>::Scalar::ZERO],
-      &[<E2 as Engine>::Scalar::ZERO],
-    );
+    let res = recursive_snark.verify(&pp);
     assert!(res.is_ok());
   }
 
@@ -902,27 +696,17 @@ mod tests {
     )
     .unwrap();
 
-    for i in 0..num_steps {
+    for _ in 0..num_steps {
       let res = recursive_snark.prove_step(&pp, &circuit_primary, &circuit_secondary);
       assert!(res.is_ok());
 
       // verify the recursive snark at each step of recursion
-      let res = recursive_snark.verify(
-        &pp,
-        i + 1,
-        &[<E1 as Engine>::Scalar::ONE],
-        &[<E2 as Engine>::Scalar::ZERO],
-      );
+      let res = recursive_snark.verify(&pp);
       assert!(res.is_ok());
     }
 
     // verify the recursive SNARK
-    let res = recursive_snark.verify(
-      &pp,
-      num_steps,
-      &[<E1 as Engine>::Scalar::ONE],
-      &[<E2 as Engine>::Scalar::ZERO],
-    );
+    let res = recursive_snark.verify(&pp);
     assert!(res.is_ok());
 
     let (zn_primary, zn_secondary) = res.unwrap();
@@ -990,12 +774,7 @@ mod tests {
     }
 
     // verify the recursive SNARK
-    let res = recursive_snark.verify(
-      &pp,
-      num_steps,
-      &[<E1 as Engine>::Scalar::ONE],
-      &[<E2 as Engine>::Scalar::ZERO],
-    );
+    let res = recursive_snark.verify(&pp);
     assert!(res.is_ok());
 
     let (zn_primary, zn_secondary) = res.unwrap();
@@ -1081,12 +860,7 @@ mod tests {
     }
 
     // verify the recursive SNARK
-    let res = recursive_snark.verify(
-      &pp,
-      num_steps,
-      &[<E1 as Engine>::Scalar::ONE],
-      &[<E2 as Engine>::Scalar::ZERO],
-    );
+    let res = recursive_snark.verify(&pp);
     assert!(res.is_ok());
 
     let (zn_primary, zn_secondary) = res.unwrap();
@@ -1247,7 +1021,7 @@ mod tests {
     }
 
     // verify the recursive SNARK
-    let res = recursive_snark.verify(&pp, num_steps, &z0_primary, &z0_secondary);
+    let res = recursive_snark.verify(&pp);
     assert!(res.is_ok());
 
     // produce the prover and verifier keys for compressed snark
@@ -1292,8 +1066,6 @@ mod tests {
       &*default_ck_hint(),
     );
 
-    let num_steps = 1;
-
     // produce a recursive SNARK
     let mut recursive_snark = RecursiveSNARK::<
       E1,
@@ -1315,12 +1087,7 @@ mod tests {
     assert!(res.is_ok());
 
     // verify the recursive SNARK
-    let res = recursive_snark.verify(
-      &pp,
-      num_steps,
-      &[<E1 as Engine>::Scalar::ONE],
-      &[<E2 as Engine>::Scalar::ZERO],
-    );
+    let res = recursive_snark.verify(&pp);
     assert!(res.is_ok());
 
     let (zn_primary, zn_secondary) = res.unwrap();
